# api/docs/swagger.yaml
definitions:
...
  auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.ClientType:
    properties:
      confirm_by:
        type: integer
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      self_recover:
        type: boolean
      self_register:
        type: boolean
      tables:
        items:
          $ref: '#/definitions/auth_service.Table'
        type: array
    type: object
...
  auth_service.Object:
    properties:
      object_id:
        type: string
      table_slug:
        type: string
    type: object
...
  auth_service.RecordPermission:
    properties:
      client_type_id:
        type: string
      delete:
        type: string
      id:
        type: string
      read:
        type: string
      role_id:
        type: string
      table_slug:
        type: string
      update:
        type: string
      write:
        type: string
    type: object
...
  auth_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
...
  auth_service.Session:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      data:
        type: string
      expires_at:
        type: string
      id:
        type: string
      integration_id:
        type: string
      ip:
        type: string
      is_changed:
        type: boolean
      project_id:
        type: string
      role_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  auth_service.Table:
    properties:
      data:
        type: string
      icon:
        type: string
      label:
        type: string
      slug:
        type: string
      view_label:
        type: string
      view_slug:
        type: string
    type: object
  auth_service.Token:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      refresh_in_seconds:
        type: integer
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  auth_service.V2LoginResponse:
    properties:
      client_platform:
        $ref: '#/definitions/auth_service.ClientPlatform'
        type: object
      client_type:
        $ref: '#/definitions/auth_service.ClientType'
        type: object
      permissions:
        items:
          $ref: '#/definitions/auth_service.RecordPermission'
        type: array
      role:
        $ref: '#/definitions/auth_service.Role'
        type: object
      sessions:
        items:
          $ref: '#/definitions/auth_service.Session'
        type: array
      token:
        $ref: '#/definitions/auth_service.Token'
        type: object
      user_found:
        type: boolean
      user_id:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
  models.Email:
    properties:
      client_type:
        type: string
      email:
        type: string
    type: object
  models.RegisterOtp:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
  models.SendCodeResponse:
    properties:
      data:
        $ref: '#/definitions/object_builder_service.V2LoginResponse'
        type: object
      sms_id:
        type: string
    type: object
...
  models.Verify:
    properties:
      data:
        $ref: '#/definitions/object_builder_service.V2LoginResponse'
        type: object
      tables:
        items:
          $ref: '#/definitions/auth_service.Object'
        type: array
    type: object
...
  object_builder_service.ClientPlatform:
    properties:
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
  object_builder_service.ClientType:
    properties:
      confirm_by:
        type: integer
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
      self_recover:
        type: boolean
      self_register:
        type: boolean
      tables:
        items:
          $ref: '#/definitions/object_builder_service.TableClientType'
        type: array
    type: object
  object_builder_service.RecordPermission:
    properties:
      client_type_id:
        type: string
      delete:
        type: string
      guid:
        type: string
      read:
        type: string
      role_id:
        type: string
      table_slug:
        type: string
      update:
        type: string
      write:
        type: string
    type: object
  object_builder_service.Role:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      guid:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
  object_builder_service.TableClientType:
    properties:
      data:
        type: string
      icon:
        type: string
      label:
        type: string
      slug:
        type: string
      view_label:
        type: string
      view_slug:
        type: string
    type: object
  object_builder_service.V2LoginResponse:
    properties:
      client_platform:
        $ref: '#/definitions/object_builder_service.ClientPlatform'
        type: object
      client_type:
        $ref: '#/definitions/object_builder_service.ClientType'
        type: object
      permissions:
        items:
          $ref: '#/definitions/object_builder_service.RecordPermission'
        type: array
      role:
        $ref: '#/definitions/object_builder_service.Role'
        type: object
      user_found:
        type: boolean
      user_id:
        type: string
    type: object
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /register-email-otp/{table_slug}:
    post:
      consumes:
      - application/json
      description: RegisterOtp
      operationId: registerEmailOtp
      parameters:
      - description: register_body
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/models.RegisterOtp'
      - description: table_slug
        in: path
        name: table_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2LoginResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: RegisterEmailOtp
      tags:
      - Email
  /register-otp/{table_slug}:
    post:
      consumes:
      - application/json
      description: RegisterOtp
      operationId: registerOtp
      parameters:
      - description: register_body
        in: body
        name: registerBody
        required: true
        schema:
          $ref: '#/definitions/models.RegisterOtp'
      - description: table_slug
        in: path
        name: table_slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2LoginResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: RegisterOtp
      tags:
      - register
...
  /send-message:
    post:
      consumes:
      - application/json
      description: Send Message to Email
      operationId: send_message_to_email
      parameters:
      - description: SendMessageToEmailRequestBody
        in: body
        name: send_message
        required: true
        schema:
          $ref: '#/definitions/models.Email'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SendCodeResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Send Message To Email
      tags:
      - Email
...
  /verify-email/{sms_id}/{otp}:
    post:
      consumes:
      - application/json
      description: Verify
      operationId: verify
      parameters:
      - description: sms_id
        in: path
        name: sms_id
        required: true
        type: string
      - description: otp
        in: path
        name: otp
        required: true
        type: string
      - description: verify_body
        in: body
        name: verifyBody
        required: true
        schema:
          $ref: '#/definitions/models.Verify'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.V2LoginResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Verify
      tags:
      - Email
...
swagger: "2.0"